HARMONY-398: Converting NavLinks to class
